ZVISION: Reorganize detection data
Migrated static detection data to detection_tables.h
and removed the need for detection.h includes. Also
edited game option descriptions.
