+++ engines/zvision/detection.cpp
@@ -24,8 +24,9 @@
 
 #include "base/plugins.h"
 
+#include "engines/advancedDetector.h"
+
 #include "zvision/zvision.h"
-#include "zvision/detection.h"
 #include "zvision/file/save_manager.h"
 #include "zvision/scripting/script_manager.h"
 
@@ -36,268 +37,30 @@
 
 namespace ZVision {
 
+struct ZVisionGameDescription {
+	ADGameDescription desc;
+	ZVisionGameId gameId;
+};
+
+ZVisionGameId ZVision::getGameId() const {
+	return _gameDescription->gameId;
+}
+Common::Language ZVision::getLanguage() const {
+	return _gameDescription->desc.language;
+}
 uint32 ZVision::getFeatures() const {
 	return _gameDescription->desc.flags;
 }
 
-Common::Language ZVision::getLanguage() const {
-	return _gameDescription->desc.language;
-}
-
 } // End of namespace ZVision
 
-static const PlainGameDescriptor zVisionGames[] = {
-	{"zvision",  "Z-Vision Game"},
-	{"znemesis", "Zork Nemesis: The Forbidden Lands"},
-	{"zgi",      "Zork: Grand Inquisitor"},
-	{0, 0}
-};
-
-namespace ZVision {
-
-#define GAMEOPTION_ORIGINAL_SAVELOAD          GUIO_GAMEOPTIONS1
-#define GAMEOPTION_DOUBLE_FPS                 GUIO_GAMEOPTIONS2
-#define GAMEOPTION_ENABLE_VENUS               GUIO_GAMEOPTIONS3
-#define GAMEOPTION_DISABLE_ANIM_WHILE_TURNING GUIO_GAMEOPTIONS4
-#define GAMEOPTION_USE_HIRES_MPEG_MOVIES      GUIO_GAMEOPTIONS5
-
-static const ZVisionGameDescription gameDescriptions[] = {
-
-	{
-		// Zork Nemesis English version
-		{
-			"znemesis",
-			0,
-			AD_ENTRY1s("CSCR.ZFS", "88226e51a205d2e50c67a5237f3bd5f2", 2397741),
-			Common::EN_ANY,
-			Common::kPlatformDOS,
-			ADGF_NO_FLAGS,
-			GUIO4(GAMEOPTION_ORIGINAL_SAVELOAD, GAMEOPTION_DOUBLE_FPS, GAMEOPTION_ENABLE_VENUS, GAMEOPTION_DISABLE_ANIM_WHILE_TURNING)
-		},
-		GID_NEMESIS
-	},
-
-	{
-		// Zork Nemesis French version
-		{
-			"znemesis",
-			0,
-			{{"CSCR.ZFS", 0, "f04113357b4748c13efcb58b4629887c", 2577873},
-			 {"NEMESIS.STR", 0, "333bcb17bbb7f57cae742fbbe44f56f3", 9219},
-			 AD_LISTEND
-			},
-			Common::FR_FRA,
-			Common::kPlatformDOS,
-			ADGF_NO_FLAGS,
-			GUIO4(GAMEOPTION_ORIGINAL_SAVELOAD, GAMEOPTION_DOUBLE_FPS, GAMEOPTION_ENABLE_VENUS, GAMEOPTION_DISABLE_ANIM_WHILE_TURNING)
-		},
-		GID_NEMESIS
-	},
-
-	{
-		// Zork Nemesis German version
-		{
-			"znemesis",
-			0,
-			{{"CSCR.ZFS", 0, "f04113357b4748c13efcb58b4629887c", 2577873},
-			 {"NEMESIS.STR", 0, "3d1a12b907751653866cffc6d4dfb331", 9505},
-			 AD_LISTEND
-			},
-			Common::DE_DEU,
-			Common::kPlatformDOS,
-			ADGF_NO_FLAGS,
-			GUIO4(GAMEOPTION_ORIGINAL_SAVELOAD, GAMEOPTION_DOUBLE_FPS, GAMEOPTION_ENABLE_VENUS, GAMEOPTION_DISABLE_ANIM_WHILE_TURNING)
-		},
-		GID_NEMESIS
-	},
-
-	{
-		// Zork Nemesis Italian version
-		{
-			"znemesis",
-			0,
-			{{"CSCR.ZFS", 0, "f04113357b4748c13efcb58b4629887c", 2577873},
-			 {"NEMESIS.STR", 0, "7c568feca8d9f9ae855c47183612c305", 9061},
-			 AD_LISTEND
-			},
-			Common::IT_ITA,
-			Common::kPlatformDOS,
-			ADGF_NO_FLAGS,
-			GUIO4(GAMEOPTION_ORIGINAL_SAVELOAD, GAMEOPTION_DOUBLE_FPS, GAMEOPTION_ENABLE_VENUS, GAMEOPTION_DISABLE_ANIM_WHILE_TURNING)
-		},
-		GID_NEMESIS
-	},
-
-	{
-		// Zork Nemesis English demo version
-		{
-			"znemesis",
-			"Demo",
-			AD_ENTRY1s("SCRIPTS.ZFS", "64f1e881394e9462305104f99513c833", 380539),
-			Common::EN_ANY,
-			Common::kPlatformWindows,
-			ADGF_DEMO,
-			GUIO4(GAMEOPTION_ORIGINAL_SAVELOAD, GAMEOPTION_DOUBLE_FPS, GAMEOPTION_ENABLE_VENUS, GAMEOPTION_DISABLE_ANIM_WHILE_TURNING)
-		},
-		GID_NEMESIS
-	},
-
-	{
-		// Zork Grand Inquisitor English CD version
-		{
-			"zgi",
-			"CD",
-			AD_ENTRY1s("SCRIPTS.ZFS", "81efd40ecc3d22531e211368b779f17f", 8336944),
-			Common::EN_ANY,
-			Common::kPlatformWindows,
-			ADGF_NO_FLAGS,
-			GUIO3(GAMEOPTION_ORIGINAL_SAVELOAD, GAMEOPTION_DOUBLE_FPS, GAMEOPTION_DISABLE_ANIM_WHILE_TURNING)
-		},
-		GID_GRANDINQUISITOR
-	},
-
-	{
-		// Zork Grand Inquisitor French CD version, reported by ulrichh on IRC
-		{
-			"zgi",
-			"CD",
-			AD_ENTRY1s("SCRIPTS.ZFS", "4d1ec4ade7ecc9ee9ec591d43ca3d213", 8338133),
-			Common::FR_FRA,
-			Common::kPlatformWindows,
-			ADGF_NO_FLAGS,
-			GUIO3(GAMEOPTION_ORIGINAL_SAVELOAD, GAMEOPTION_DOUBLE_FPS, GAMEOPTION_DISABLE_ANIM_WHILE_TURNING)
-		},
-		GID_GRANDINQUISITOR
-	},
-
-	{
-		// Zork Grand Inquisitor German CD version, reported by breit in bug #6760
-		{
-			"zgi",
-			"CD",
-			AD_ENTRY1s("SCRIPTS.ZFS", "b7ac7e331b9b7f884590b0b325b560c8", 8338133),
-			Common::DE_DEU,
-			Common::kPlatformWindows,
-			ADGF_NO_FLAGS,
-			GUIO3(GAMEOPTION_ORIGINAL_SAVELOAD, GAMEOPTION_DOUBLE_FPS, GAMEOPTION_DISABLE_ANIM_WHILE_TURNING)
-		},
-		GID_GRANDINQUISITOR
-	},
-
-	{
-		// Zork Grand Inquisitor Spanish CD version, reported by dianiu in bug #6764
-		{
-			"zgi",
-			"CD",
-			AD_ENTRY1s("SCRIPTS.ZFS", "5cdc4b99c1134053af135aae71326fd1", 8338141),
-			Common::ES_ESP,
-			Common::kPlatformWindows,
-			ADGF_NO_FLAGS,
-			GUIO3(GAMEOPTION_ORIGINAL_SAVELOAD, GAMEOPTION_DOUBLE_FPS, GAMEOPTION_DISABLE_ANIM_WHILE_TURNING)
-		},
-		GID_GRANDINQUISITOR
-	},
-
-	{
-		// Zork Grand Inquisitor English DVD version
-		{
-			"zgi",
-			"DVD",
-			AD_ENTRY1s("SCRIPTS.ZFS", "03157a3399513bfaaf8dc6d5ab798b36", 8433326),
-			Common::EN_ANY,
-			Common::kPlatformWindows,
-			ADGF_NO_FLAGS,
-			GUIO4(GAMEOPTION_ORIGINAL_SAVELOAD, GAMEOPTION_DOUBLE_FPS, GAMEOPTION_DISABLE_ANIM_WHILE_TURNING, GAMEOPTION_USE_HIRES_MPEG_MOVIES)
-		},
-		GID_GRANDINQUISITOR
-	},
-
-	{
-		// Zork Grand Inquisitor English demo version
-		{
-			"zgi",
-			"Demo",
-			AD_ENTRY1s("SCRIPTS.ZFS", "71a2494fd2fb999347deb13401e9b998", 304239),
-			Common::EN_ANY,
-			Common::kPlatformWindows,
-			ADGF_DEMO,
-			GUIO3(GAMEOPTION_ORIGINAL_SAVELOAD, GAMEOPTION_DOUBLE_FPS, GAMEOPTION_DISABLE_ANIM_WHILE_TURNING)
-		},
-		GID_GRANDINQUISITOR
-	},
-
-	{
-		AD_TABLE_END_MARKER,
-		GID_NONE
-	}
-};
-
-} // End of namespace ZVision
-
-static const char *directoryGlobs[] = {
-	"znemscr",
-	0
-};
-
-static const ADExtraGuiOptionsMap optionsList[] = {
-	{
-		GAMEOPTION_ORIGINAL_SAVELOAD,
-		{
-			_s("Use original save/load screens"),
-			_s("Use the original save/load screens, instead of the ScummVM ones"),
-			"originalsaveload",
-			false
-		}
-	},
-
-	{
-		GAMEOPTION_DOUBLE_FPS,
-		{
-			_s("Double FPS"),
-			_s("Increase game FPS from 30 to 60"),
-			"doublefps",
-			false
-		}
-	},
-
-	{
-		GAMEOPTION_ENABLE_VENUS,
-		{
-			_s("Enable Venus"),
-			_s("Enable the Venus help system"),
-			"venusenabled",
-			true
-		}
-	},
-
-	{
-		GAMEOPTION_DISABLE_ANIM_WHILE_TURNING,
-		{
-			_s("Disable animation while turning"),
-			_s("Disable animation while turning in panoramic mode"),
-			"noanimwhileturning",
-			false
-		}
-	},
-
-	{
-		GAMEOPTION_USE_HIRES_MPEG_MOVIES,
-		{
-			_s("Use the hires MPEG movies"),
-			_s("Use the hires MPEG movies of the DVD version, instead of the lowres AVI ones"),
-			"mpegmovies",
-			true
-		}
-	},
-
-	AD_EXTRA_GUI_OPTIONS_TERMINATOR
-};
+#include "zvision/detection_tables.h"
 
 class ZVisionMetaEngine : public AdvancedMetaEngine {
 public:
-	ZVisionMetaEngine() : AdvancedMetaEngine(ZVision::gameDescriptions, sizeof(ZVision::ZVisionGameDescription), zVisionGames, optionsList) {
+	ZVisionMetaEngine() : AdvancedMetaEngine(ZVision::gameDescriptions, sizeof(ZVision::ZVisionGameDescription), ZVision::zVisionGames, ZVision::optionsList) {
 		_maxScanDepth = 2;
-		_directoryGlobs = directoryGlobs;
+		_directoryGlobs = ZVision::directoryGlobs;
 		_singleid = "zvision";
 	}
