+++ engines/zvision/text/string_manager.h
@@ -23,7 +23,6 @@
 #ifndef ZVISION_STRING_MANAGER_H
 #define ZVISION_STRING_MANAGER_H
 
-#include "zvision/detection.h"
 #include "zvision/text/truetype_font.h"
 
 namespace Graphics {
